refactor(webui): split settings backend by domain
